Rename tools
The module tools should be hidden and not appear in the documentation pages as it only contains auxiliary functions
